requirements.txt
Babel==1.3
Flask==0.10
futures==3.0.5
gunicorn==19.6.0
image==1.5.5
imageio==1.6
itsdangerous==0.24
Jinja2==2.8
Pillow==3.2.0
certifi>=0.0.8
chardet>=1.0.1
distribute>=0.6.24
Werkzeug==0.9
requests>=0.11.1
greenlet>=0.4.10
gevent>=1.1.2
